+++ script/ability/effects.ts
@@ -169,6 +169,19 @@ const effects = {
       mpRegenV: 1,
     },
   } as EffectObject,
+  boss_effect: {
+    id: "boss_effect",
+    icon: "gfx/icons/brute.png",
+    duration: 0,
+    isInfinite: true,
+    type: "heal",
+    buffDebuff: "buff",
+    modifiers: {
+      stunResistanceV: 20,
+      bleedResistanceV: 20,
+      poisonResistanceV: 20,
+    },
+  } as EffectObject,
   troll_regen: {
     id: "troll_regen",
     icon: "gfx/status/great-heart.png",