+++ script/character/character.ts
@@ -128,8 +128,8 @@ class Character {
       this.updateAllModifiers();
       const resistances = { ...this.resistances };
       Object.entries(resistances).map(([key, value]) => {
-        let modifier = this.allModifiers[key + "_resistanceP"] ?? 1;
-        let boost = this.allModifiers[key + "_resistanceV"] ?? 0;
+        let modifier = this.allModifiers[key + "ResistanceP"] ?? 1;
+        let boost = this.allModifiers[key + "ResistanceV"] ?? 0;
         boost += ((this.level - 1) | 0) * 0.2; // Level resistance boost
         resistances[key] = Math.floor((value + boost) * modifier);
       });
@@ -268,6 +268,20 @@ class Character {
     this.addStatus = (status: Effect, user: Player | Enemy, key: string): void => {
       const index = this.statuses.findIndex((s: any) => s.id === status.id);
       const effect = new Effect(status);
+      const resistance = this.getResistances()[effect.type];
+      // If the effect is stun, check if it's resisted
+      if (effect.type === "stun") {
+        if (Math.random() * 120 < resistance) {
+          if (this instanceof Player) {
+            createDroppingText("Stun resisted!", tools, "resist");
+          } else {
+            createDroppingText("Stun resisted!", this.card.main, "resist");
+          }
+          return;
+        } else {
+          effect.duration = Math.round(effect.duration * (1 - resistance / 180));
+        }
+      }
       if (index === -1) {
         effect.init(user.allModifiers?.[key]?.["effect_" + status.id]);
         effect.lasts = effect.duration;